Add Guzzle/ClientFactory and 2.1.0
This factory can be used to get a guzzle client
with a rpeset user agent, cookie jar and also
handler for retrying mediawiki requests.

This factory is needed by the wikibase-commands
in order to be able to inject a client into
MediawikiApi for serveral mediawiki sites to perform
async calls accross multiple sites.

Otherwise async will not correctly work!